services/agent-api/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-cloud-storage
numpy
Pillow
google-genai
